pyproject.toml
[project]
name = "Xerolysis"
description = "Servicio de blotting: Ξηρο-Λύσις (Xero-Lysis) - Solución de Secado"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["django-cotton>=1.5.1", "django>=5.1.4", "python-dotenv>=1.0.1"]

[dependency-groups]
dev = ["djlint>=1.36.4", "pytest>=8.3.4", "pytest-django>=4.9.0"]

[tool.ruff]
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
ignore = ["F401"]
extend-select = ["F", "I", "N", "A"]

[tool.ruff.format]
quote-style = "single"

[tool.djlint]
profile = "django"
ignore = "H021,H030,H031,T002"
blank_line_before_tag = "block"
blank_line_after_tag = "load,extends,endblock"
format_attribute_template_tags = true
line_break_after_multiline_tag = false
max_line_length = 120
max_attribute_length = 80
custom_html = "c-\\w+"
# close_void_tags = true
# format_css = true
# format_js = true
# no_set_formatting = true
# no_function_formatting = true
